feat(cli): Added plugin loading
Currently the only plugin loader that is implemented is NPM loading
which uses NPM style imports/requires to load modules as plugins.
Eventually others might be added such as WASM support which would be
safer if untrusted plugins need to be included or support for loading
plugins more dynamically without requiring modifications to the
environment.