feat(micro-frontend): Added micro frontend library
This provides a simple implementation of a micro frontend architecture
which can be used to distribute the development of different parts of a
UI to different groups to be developed and deployed independently.

While everything will ultimately be bundled together with this app for
now, this makes it much easier to dynamically build out different parts
of the UI and facilitates switching from React to other frameworks if
needed.